+++ src/components/Sidebar/SidebarFooter/SidebarFooter.tsx
@@ -1,13 +1,39 @@
+// FRAMEWORK ===========================================================================================================
 import React from "react";
 
+// FICTOAN =============================================================================================================
 import { Element } from "../../Element/Element";
-import { CommonAndHTMLProps } from "../../Element/constants";
 
+// STYLES ==============================================================================================================
 import "./sidebar-footer.css";
 
-export type SidebarFooterElementType = HTMLDivElement;
-export type SidebarFooterProps = CommonAndHTMLProps<SidebarFooterElementType>;
+// TYPES ===============================================================================================================
+import { CommonAndHTMLProps } from "../../Element/constants";
 
-export const SidebarFooter = React.forwardRef((props: SidebarFooterProps, ref: React.Ref<SidebarFooterElementType>) => {
-    return <Element<SidebarFooterElementType> as="footer" data-sidebar-footer ref={ref} {...props} />;
-});
+// prettier-ignore
+export interface SidebarFooterCustomProps {
+    isSticky ? : boolean;
+}
+
+export type SidebarFooterElementType = HTMLDivElement;
+export type SidebarFooterNewProps = Omit<CommonAndHTMLProps<SidebarFooterElementType>, keyof SidebarFooterCustomProps>
+    & SidebarFooterCustomProps;
+
+export const SidebarFooter = React.forwardRef(
+    ({ isSticky, ...props }: SidebarFooterNewProps, ref: React.Ref<SidebarFooterElementType>) => {
+        let classNames = [];
+
+        if (isSticky) {
+            classNames.push("is-sticky");
+        }
+
+        return (
+            <Element<SidebarFooterElementType>
+                as="footer"
+                data-sidebar-footer
+                ref={ref}
+                classNames={classNames}
+                {...props}
+            />
+        );
+    });